Changed field initialInterval by startAlignedInterval
The target interval is calculated based on the emissions per second, so the initial interval should also be calculated based on it. The correct way would be, then, to have it being recalculated when the emissions per seconds change, maybe based on some kind of percent given by the user. However, this normally is only needed to align the interval with the start of the emission (so a boolean is more suitable and easy to understand).